refactor: extract invoice credit line item conversion
